Customer Order Number now goes to comment field on invoice
Invoice no longer has a separate customer order number field. Instead, the
CON from sale is placed in the comment (Terms and Conditions) field, so that
it gets transferred when submitting invoices to external accounting software.

The field is not yet deleted from invoice, just removed from the form view.